# openapi.yaml
openapi: 3.0.0
info:
  title: Project Spark Blueprint API
  version: v1
  description: API for Project Spark Blueprint application
servers:
  - url: http://localhost:5000/v1 # Assuming API v1 prefix is handled by Flask app
    description: Local development server
paths:
...
  /comments/project/{project_id}/comments:
    get:
      summary: List comments for a project
      operationId: listProjectComments
      tags:
        - Comments
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          description: ID of the project
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A list of comments for the project
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentOutput' # Updated to CommentOutput
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      summary: Create a comment for a project
      operationId: createProjectComment
      tags:
        - Comments
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          description: ID of the project
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentOutput' # Updated to CommentOutput
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /comments/task/{task_id}/comments:
    get:
      summary: List comments for a task
      operationId: listTaskComments
      tags:
        - Comments
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          description: ID of the task
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A list of comments for the task
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentOutput' # Updated to CommentOutput
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      summary: Create a comment for a task
      operationId: createTaskComment
      tags:
        - Comments
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          description: ID of the task
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentOutput' # Updated to CommentOutput
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /comments/{comment_id}:
    get:
      summary: Get a comment by ID
      operationId: getCommentById
      tags:
        - Comments
      security:
        - bearerAuth: []
      parameters:
        - name: comment_id
          in: path
          required: true
          description: ID of the comment
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Comment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentOutput' # Updated to CommentOutput
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      summary: Update a comment by ID
      operationId: updateComment
      tags:
        - Comments
      security:
        - bearerAuth: []
      parameters:
        - name: comment_id
          in: path
          required: true
          description: ID of the comment to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentUpdateInput' # Updated to CommentUpdateInput
      responses:
        '200':
          description: Comment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentOutput' # Updated to CommentOutput
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      summary: Delete a comment by ID
      operationId: deleteComment
      tags:
        - Comments
      security:
        - bearerAuth: []
      parameters:
        - name: comment_id
          in: path
          required: true
          description: ID of the comment to delete
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Comment deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
components:
  schemas:
...
    ErrorResponse: # Generic error response
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
    UserSimpleOutput: # Simplified user for nested representations
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        email:
          type: string
          format: email
...
    CommentBase:
      type: object
      properties:
        text_content:
          type: string
          description: The content of the comment
        project_id:
          type: string
          format: uuid
          nullable: true
          description: ID of the project this comment belongs to (if any)
        task_id:
          type: string
          format: uuid
          nullable: true
          description: ID of the task this comment belongs to (if any)
        parent_comment_id:
          type: string
          format: uuid
          nullable: true
          description: ID of the parent comment if this is a reply
    CommentInput: # Schema for creating a new comment
      allOf:
        - $ref: '#/components/schemas/CommentBase' # text_content, project_id, task_id, parent_comment_id
      type: object
      required:
        - text_content
      # project_id or task_id will be part of path for creation usually, but can be in body if a generic endpoint is used
      # For our specific routes (/project/{id}/comments and /task/{id}/comments), project_id/task_id are from path
      # So the input schema for those POST bodies might only need text_content and parent_comment_id.
      # Let's simplify CommentInput to reflect this for the body of POST requests
      # Simplified CommentInput for POST bodies where context (project/task) is from path:
...
    CommentUpdateInput: # Schema for updating a comment
      type: object
      properties:
        text_content:
          type: string
          description: New content for the comment
          minLength: 1
          nullable: true # Allow sending null to clear? Or just make it optional.
                        # For now, assume if sent, it must be non-empty. If not sent, no change.
    CommentOutput: # Schema for comment responses
      type: object
      properties:
        id:
          type: string
          format: uuid
        text_content:
          type: string
        author:
          $ref: '#/components/schemas/UserSimpleOutput' # Reusing the existing UserSimpleOutput
        project_id:
          type: string
          format: uuid
          nullable: true
        task_id:
          type: string
          format: uuid
          nullable: true
        parent_comment_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        replies:
          type: array
          items:
            $ref: '#/components/schemas/CommentOutput' # Recursive definition for replies
      required:
        - id
        - text_content
        - author
        - created_at
        - updated_at
        - replies
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # Adding common responses to avoid repetition
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFoundError:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequestError:
      description: The request payload is malformed or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - bearerAuth: [] 
# The global security definition means all endpoints require bearerAuth by default.
# For public endpoints like /auth/login and /auth/register, 
# you need to override this at the operation level with `security: []`
# Let's adjust paths for /auth/login and /auth/register to explicitly have no security

# Adjusting auth paths to have no global security
# This should be done by finding these paths in the existing openapi.yaml and adding 'security: []'
# For this edit, I will assume the edit tool handles merging this correctly.
# If not, a separate edit focusing only on auth paths might be needed.

# Path adjustments for auth (example, needs to be merged with existing auth paths)
# /auth/register:
#   post:
#     security: [] 
#     # ... rest of the definition
# /auth/login:
#   post:
#     security: []
#     # ... rest of the definition
# /auth/refresh: # Refresh token might or might not need Bearer auth, depending on flow.
               # Assuming refresh token itself is the auth for this one, also no Bearer.
#   post:
#     security: [] 
#     # ... rest of the definition 
...